// assets/ts/admin/interfaces/ResizeOptions.ts
/*
 * This file is part of the NumberNine package.
 *
 * (c) William Arin <williamarin.dev@gmail.com>
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */

export default interface ResizeOptions {
    enabled?: boolean;
    width: number;
    height: number;
    quality?: number;
    mimeType?: string;
}